refactor(api): move paginate util to src
