+++ api/tests/shared/unit/infrastructure/utils/paginate_test.js
@@ -1,5 +1,5 @@
-import { paginate } from '../../../../lib/infrastructure/utils/paginate.js';
-import { expect } from '../../../test-helper.js';
+import { paginate } from '../../../../../src/shared/infrastructure/utils/paginate.js';
+import { expect } from '../../../../test-helper.js';
 
 describe('Unit | Utils | paginate', function () {
   describe('#paginate', function () {
